StockSearchService/requirements.txt
pandas~=2.2.1
mysql-connector-python~=8.3.0
Flask~=3.0.2
Flask-CORS ~= 4.0.0
python-dotenv~=1.0.1